Implement EnsembleEvaluatorAsync, which will be activated when SchedulerFeature is enabled.
A new asyncio.loop is set in base_run_mode to run the entire EEAsync.
This follows a similar pattern to Scheduler; ie. create a set of
background tasks responsible for running the ee server, processing
events, sending snapshots updates, etc.
NB! It does not use BatchingDispatcher abd it adds MonitorAsync for running the tests as the current one does not work with eeasync.